+++ CHANGES.md
@@ -1,5 +1,7 @@
 # Changes
 
+* Fix decode uncomple array data
+
 ## [0.4.0] - 2023-01-04
 
 * 0.4 Release